+++ Reprise.SampleApi/Endpoints/Users/DeleteUserEndpoint.cs
@@ -4,6 +4,8 @@
     public class DeleteUserEndpoint
     {
         [Delete("/users/{id}")]
+        [Produces(StatusCodes.Status204NoContent)]
+        [Produces(StatusCodes.Status404NotFound)]
         public static IResult Handle(int id, DataContext context)
         {
             var index = context.Users.FindIndex(x => x.Id == id);