+++ Reprise.SampleApi/Endpoints/Users/GetAllUsersEndpoint.cs
@@ -4,6 +4,7 @@
     public class GetAllUsersEndpoint
     {
         [Get("/users")]
+        [Produces(StatusCodes.Status200OK, typeof(IEnumerable<User>))]
         public static IEnumerable<User> Handle(DataContext context) => context.Users;
     }
 }