+++ Reprise.SampleApi/Endpoints/Users/UpdateUserEndpoint.cs
@@ -4,6 +4,9 @@
     public class UpdateUserEndpoint
     {
         [Put("/users/{id}")]
+        [Produces(StatusCodes.Status204NoContent)]
+        [Produces(StatusCodes.Status404NotFound)]
+        [ProducesBadRequest]
         public static IResult Handle(int id, UserDto userDto, DataContext context)
         {
             var user = context.Users.FirstOrDefault(u => u.Id == id);