+++ Reprise.SampleApi/Usings.cs
@@ -1,4 +1,5 @@
 global using FluentValidation;
 global using Microsoft.AspNetCore.Authorization;
 global using Reprise.SampleApi.Data;
+global using Reprise.SampleApi.ErrorHandling;
 global using Reprise.SampleApi.Filters;