+++ Reprise/Features/Authorization.cs
@@ -16,7 +16,7 @@
         }
     }
 
-    internal class AuthorizationProcessor : IRouteHandlerBuilderProcessor
+    internal sealed class AuthorizationProcessor : IRouteHandlerBuilderProcessor
     {
         public void Process(RouteHandlerBuilder builder, MethodInfo handlerInfo, EndpointOptions options, string route)
         {